navitationBar y Card integrada a Home
