+++ src/containers/home/ContainerHome.jsx
@@ -6,7 +6,8 @@ import Categories from "../../components/categories/Categories";
 import Featured from "../../components/featured/Featured";
 import tw from "twrnc";
 
-const ContainerHome = ({ navigation }) => {
+const ContainerHome = ({ navigation }) =>
+{
   return (
     <View style={tw`flex m-2 p-4`}>
       <UserName />
